Move loadExercise & refreshExercise
